generator building

+++ console/folder/migrate/migration/1480011619_building.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `building`
         (
-            `building_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
+            `building_id` TINYINT(1) PRIMARY KEY AUTO_INCREMENT,
             `building_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `building` (`building_name`)
